use long instead of int in initMersenneTwister()
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rmetrics@5087 5d4a7380-5c0c-0410-b446-d4a5018c0119

+++ pkg/randtoolbox/src/mt19937ar.c
@@ -183,7 +183,7 @@ double MT2002_genrand_res53(void)
 #include "runifInterface.h"
 int initialization, resolution;
 
-void initMersenneTwister(int *type, int *nseed, unsigned int *iseed, unsigned int *state)
+void initMersenneTwister(int *type, int *nseed, unsigned long *iseed, unsigned int *state)
 {
 	int i;
 	if (*type == 1)
